+++ app/code/ThirdParty/BlogArticle/Api/Data/TagInterface.php
@@ -6,6 +6,7 @@ interface TagInterface
     public const TAG_ID = 'tag_id';
     public const NAME = 'name';
     public const URL_KEY = 'url_key';
+    public const DESCRIPTION = 'description';
     public const IS_ACTIVE = 'is_active';
     public const CREATION_TIME = 'creation_time';
     public const UPDATE_TIME = 'update_time';
@@ -22,6 +23,10 @@ interface TagInterface
     public function getUrlKey();
     /** @param string $urlKey @return $this */
     public function setUrlKey($urlKey);
+    /** @return string|null */
+    public function getDescription();
+    /** @param string|null $description @return $this */
+    public function setDescription($description);
     /** @return int|null */
     public function getIsActive();
     /** @param int $isActive @return $this */